Make key auto-repeat detection possible for engines.
If auto-repeat is enabled for a key, the sequence will normally be:

press, release, press, release... press, relase

Which means it's impossible for an engine to detect auto-repeat events
(and for example filter them out to detect when a key is really
released).

Using XkbSetDetectableAutoRepeat change the sequence to:

press, press... release

So it's now possible to filter those auto-repeat events.